Dynare++: in printf statements with size_t, remove %zd modifier introduced in 757d8c36e91b7ca73e034e8a4d4a0827ef0d5416 and replace it by %lu as a workaround. The %zd modifier was introduced in C99 standard, but is not implemented by MinGW.

+++ dynare++/integ/src/quadrature-points.cpp
@@ -141,7 +141,7 @@ int main(int argc, char** argv)
 			delete *it;
 		points.erase(new_end, points.end());
 
-		printf("Duplicit nodes removed:   %zd\n", sq.numEvals(level)-points.size());
+		printf("Duplicit nodes removed:   %lu\n", (unsigned long) (sq.numEvals(level)-points.size()));
 
 		// calculate weights and mass
 		double mass = 0.0;
